+++ src/components/Error-modal/ErrorModal.css
@@ -6,8 +6,10 @@
   border-radius: 4px;
   box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.15);
   position: absolute;
-  left: 531px;
-  top: 226px;
+  position: absolute;
+  right: 40px;
+  top: 20.93%;
+  /* transform: translate(-84%, -21%); */
 }
 .alert {
   display: flex;
